+++ documents/docs/zh-CN/changelog.md
@@ -28,6 +28,8 @@
 
 [新增] 子表 新增子表事件：`toolbar/edit/checkbox`
 
+[新增] 拖拽行 [指定列导出](/#/zh-CN/component/basic/row-drag)
+
 ### **1.3.17** <small>`2019-09-08`</small>
 
 [修复] 筛选 修复 url+不分页时，筛选没有恢复url，造成 reload 时，没有重新请求数据的问题，同时此问题也会引起 checkbox 的多重绑定，导致报错。